% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleiss_data.R
\docType{data}
\name{fleiss_data}
\alias{fleiss_data}
\title{Fleiss data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 4 columns.
}
\usage{
fleiss_data
}
\description{
Fleiss dataset from \code{epitools} package examples, with event counts and population per age group in tidy format.
}
\keyword{datasets}
