% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer_std_pop.R
\docType{data}
\name{seer_std_pop}
\alias{seer_std_pop}
\title{Standard population reference table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\usage{
seer_std_pop
}
\description{
This table present standard population reference for age groups from SEER*Stat WHO adjusted proportions.
}
\keyword{datasets}
