% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selvin_data_1940.R
\docType{data}
\name{selvin_data_1940}
\alias{selvin_data_1940}
\title{Selvin data, 1940}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
selvin_data_1940
}
\description{
Selvin dataset from \code{epitools} package examples for 1940, with event counts and population per age group in tidy format.
}
\keyword{datasets}
