% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_pop.R
\docType{data}
\name{std_pop}
\alias{std_pop}
\title{Standard population reference table}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\usage{
std_pop
}
\description{
This table present standard population reference for age groups from the World Health Organization (WHO) and the SEER*Stat adjusted proportions.
}
\keyword{datasets}
