% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_std_pop.R
\docType{data}
\name{who_std_pop}
\alias{who_std_pop}
\title{Standard population reference table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 1 columns.
}
\usage{
who_std_pop
}
\description{
This table present standard population reference for age groups from the World Health Organization (WHO).
}
\keyword{datasets}
